refact: Separated hooks & styles from EditAssignmentForm, corrected import & declaration on Assignments.tsx
